fix:sync optional default false chore:move to with-storage-sync folder
