+++ projects/ngrx-extension/src/public-api.ts
@@ -1,1 +1,1 @@
-export * from './lib/with-storage-sync';
+export * from './lib/with-storage-sync/with-storage-sync';
